bigdata-weather/src/test/java/com/xebia/university/bigdata/weather/KnmiDailyLineParserTest.java
package com.xebia.university.bigdata.weather;

import static org.junit.Assert.*;

import org.junit.Before;
import org.junit.Test;

import com.xebia.university.bigdata.weather.KnmiLineParser.KnmiLineType;

public class KnmiDailyLineParserTest {

    private KnmiDailyLineParser parser;

    @Before
    public void setUp() {
        this.parser = new KnmiDailyLineParser();
    }

    @Test
    public void shouldParseMeasurementLine() {
        String line = "  240,20100106,  186,   32,   34,   60,   17,   20,    8,   80,   17,  -28,  -66,   24,    3,   15,  -80,   24,   55,   70,  449,   29,   28,   13,   16,10036,10057,   23,10022,    4,    3,   16,   58,   14,    5,   95,   98,    4,   87,   13,    4";
        KnmiLineType result = this.parser.parse(line);
        assertEquals(KnmiLineType.DATA, result);
        assertEquals("240", parser.getStation());
        assertEquals("20100106", parser.getDate());
        assertEquals(28, parser.getPrecipitation());
    }

}
